+++ element_animal/surgery.py
@@ -16,7 +16,7 @@ def activate(
     *,
     create_schema: bool = True,
     create_tables: bool = True,
-    linking_module: bool = True,
+    linking_module: str = None,
 ):
     """Activate this schema.
 
@@ -27,8 +27,8 @@ def activate(
                             database if it does not yet exist.
         create_tables (bool): when True (default), create tables in the
                             database if they do not yet exist.
-        linking_module (bool): a module name or a module containing the
-        required dependencies to activate the `subject` element:
+        linking_module (str): A module name or a module containing the required
+            dependencies to activate the `surgery` module.
 
     Dependencies:
     Upstream tables:
@@ -54,7 +54,7 @@ def activate(
         surgery_schema_name,
         create_schema=create_schema,
         create_tables=create_tables,
-        add_objects=linking_module.__dict__,
+        add_objects=_linking_module.__dict__,
     )