+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
+## [0.2.1] - 2023-11-28
++ Add - `Species` table in `subject` schema
+
 ## [0.2.0] - 2023-08-23
 
 + Add - `injection` schema
@@ -63,6 +66,7 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 + Add - `subject` schema
 + Add - `genotyping` schema
 
+[0.2.1]: https://github.com/datajoint/element-animal/releases/tag/0.2.1
 [0.2.0]: https://github.com/datajoint/element-animal/releases/tag/0.2.0
 [0.1.8]: https://github.com/datajoint/element-animal/releases/tag/0.1.8
 [0.1.7]: https://github.com/datajoint/element-animal/releases/tag/0.1.7